moved constants and xml helpers to seperate files
