+++ sdcparser/src/main/java/com/sdc/parser/Constants.java
@@ -1,0 +1,17 @@
+package com.sdc.parser;
+
+public final class Constants {
+	private Constants() {
+		// restrict instantiation
+	}
+
+	public static final String INTEGER = "integer";
+	public static final String DECIMAL = "decimal";
+	public static final String STRING = "string";
+	public static final String BOOLEAN = "boolean";
+	public static final String DATE = "date";
+	public static final String DATETIME = "dateTime";
+
+//	public static final String provenanceHeader = "{\"resourceType\": \"Provenance\",\"meta\": {\"versionId\": \"1\",\"lastUpdated\": \"2020-08-31T20:44:24.994+00:00\"},\"recorded\": \"2020-05-14T13:44:24.1703291-07:00\",\"agent\": [{\"type\": {\"text\": \"Joel and Alex testing\"}}]}";
+
+}
